+++ packages/tools/pluggable-widgets-tools/configs/rollup.config.native.js
@@ -1,4 +1,6 @@
-import { existsSync, mkdirSync } from "fs";
+/* eslint-disable @typescript-eslint/explicit-function-return-type */
+
+import { existsSync } from "fs";
 import { join, relative } from "path";
 import { getBabelInputPlugin, getBabelOutputPlugin } from "@rollup/plugin-babel";
 import commonjs from "@rollup/plugin-commonjs";
@@ -12,10 +14,8 @@ import { red, yellow, blue } from "ansi-colors";
 import loadConfigFile from "rollup/dist/loadConfigFile";
 import clear from "rollup-plugin-clear";
 import command from "rollup-plugin-command";
-import license from "rollup-plugin-license";
 import { terser } from "rollup-plugin-terser";
 import { cp } from "shelljs";
-import { zip } from "zip-a-folder";
 import { widgetTyping } from "./rollup-plugin-widget-typing";
 import { collectDependencies } from "./rollup-plugin-collect-dependencies";
 import {
@@ -28,6 +28,7 @@ import {
     widgetPackage,
     widgetVersion
 } from "./shared";
+import { copyLicenseFile, createMpkFile, licenseCustomTemplate } from "./helpers/rollup-helper";
 
 const outDir = join(sourcePath, "/dist/tmp/widgets/");
 const outWidgetFile = join(widgetPackage.replace(/\./g, "/"), widgetName.toLowerCase(), `${widgetName}`);
@@ -36,6 +37,7 @@ const mpkFile = join(mpkDir, process.env.MPKOUTPUT ? process.env.MPKOUTPUT : `${
 
 export default async args => {
     const production = Boolean(args.configProduction);
+
     if (!production && projectPath) {
         console.info(blue(`Project Path: ${projectPath}`));
     }
@@ -62,7 +64,28 @@ export default async args => {
                 }),
                 ...(i === 0 ? getClientComponentPlugins() : []),
                 json(),
-                collectDependencies({ outputDir: outDir, onlyNative: true, widgetName }),
+                collectDependencies({
+                    outputDir: outDir,
+                    onlyNative: true,
+                    widgetName,
+                    ...(production && i === 0
+                        ? {
+                              licenseOptions: {
+                                  thirdParty: {
+                                      output: [
+                                          {
+                                              file: join(outDir, "dependencies.txt")
+                                          },
+                                          {
+                                              file: join(outDir, "dependencies.json"),
+                                              template: licenseCustomTemplate
+                                          }
+                                      ]
+                                  }
+                              }
+                          }
+                        : null)
+                }),
                 ...getCommonPlugins({
                     sourceMaps: false,
                     extensions: [`.${os}.js`, ".native.js", ".js", ".jsx", ".ts", ".tsx"],
@@ -170,36 +193,22 @@ export default async args => {
                 : null,
             image(),
             production ? terser({ mangle: false }) : null,
-            config.licenses
-                ? license({
-                      thirdParty: {
-                          includePrivate: true,
-                          output: {
-                              file: join(outDir, "dependencies.txt")
-                          }
-                      }
-                  })
-                : null,
             // We need to create .mpk and copy results to test project after bundling is finished.
             // In case of a regular build is it is on `writeBundle` of the last config we define
             // (since rollup processes configs sequentially). But in watch mode rollup re-bundles only
             // configs affected by a change => we cannot know in advance which one will be "the last".
             // So we run the same logic for all configs, letting the last one win.
             command([
-                async () => {
-                    mkdirSync(mpkDir, { recursive: true });
-                    await zip(outDir, mpkFile);
-                    if (!production && projectPath) {
-                        const widgetsPath = join(projectPath, "widgets");
-                        const deploymentPath = join(projectPath, `deployment/native/widgets`);
-                        // Create folder if they do not exists or directories were cleaned
-                        mkdirSync(widgetsPath, { recursive: true });
-                        mkdirSync(deploymentPath, { recursive: true });
-                        // Copy files to deployment and widgets folder
-                        cp("-r", join(outDir, "*"), deploymentPath);
-                        cp(mpkFile, widgetsPath);
-                    }
-                }
+                async () => config.licenses && copyLicenseFile(sourcePath, outDir),
+                async () =>
+                    createMpkFile({
+                        mpkDir,
+                        mpkFile,
+                        widgetTmpDir: outDir,
+                        isProduction: production,
+                        mxProjectPath: projectPath,
+                        deploymentPath: "deployment/native/widgets"
+                    })
             ])
         ];
     }